it averages but works slowly

+++ math module/math.h
@@ -2,4 +2,4 @@
 #include <vector>
 
 std::vector<double> generatePoints(double distance, int num_points);
-std::vector<double> rawDataToVolts(const std::vector<int16_t>& raw_data, double v_div);
+std::vector<double> rawDataToVolts(const std::vector<int32_t>& raw_data, double v_div);
